+++ lib/generators/react_on_rails/generator_helper.rb
@@ -1,8 +1,13 @@
 # frozen_string_literal: true
 
 require "rainbow"
+require_relative "option_helper"
 
 module GeneratorHelper
+  def src_file_exists?(file)
+    File.exist?(File.join("templates", file))
+  end
+
   # Takes a relative path from the destination root, such as `.gitignore` or `app/assets/javascripts/application.js`
   def dest_file_exists?(file)
     dest_file = File.join(destination_root, file)
@@ -22,6 +27,14 @@ Please add the following content to your #{file} file:
 #{data}
     MSG
     # rubocop:enable Layout/IndentHeredoc
+  end
+
+  def create_client_directories(*dir_names)
+    dir_names.each do |dir_name|
+      dest_dir = "client/app/bundles/HelloWorld/#{dir_name}"
+      empty_directory \
+        convert_filename_to_use_example_page_name(dest_dir)
+    end
   end
 
   def empty_directory_with_keep_file(destination, config = {})
@@ -53,5 +66,15 @@ Please add the following content to your #{file} file:
     parent_directories = destination_path.dirname
     empty_directory(parent_directories) unless dest_dir_exists?(parent_directories)
     copy_file source_file, destination_file
+  end
+
+  def copy_or_template(src_file, dest_file)
+    if src_file_exists?(src_file)
+      copy_file(src_file, dest_file)
+      return
+    end
+
+    src_template = "#{src_file}.tt"
+    template(src_template, dest_file)
   end
 end