+++ lib/generators/react_on_rails/react_with_redux_generator.rb
@@ -1,16 +1,19 @@
 # frozen_string_literal: true
 
 require "rails/generators"
+require_relative "generator_helper"
 
 module ReactOnRails
   module Generators
     class ReactWithReduxGenerator < Rails::Generators::Base
+      include GeneratorHelper
+      include OptionHelper
       Rails::Generators.hide_namespace(namespace)
       source_root(File.expand_path("../templates", __FILE__))
+      define_name_option
 
       def create_redux_directories
-        dirs = %w[actions constants reducers store]
-        dirs.each { |name| empty_directory("client/app/bundles/HelloWorld/#{name}") }
+        create_client_directories "actions", "constants", "reducers", "store"
       end
 
       def copy_base_redux_files
@@ -22,7 +25,7 @@ module ReactOnRails
            client/app/bundles/HelloWorld/reducers/helloWorldReducer.jsx
            client/app/bundles/HelloWorld/store/helloWorldStore.jsx
            client/app/bundles/HelloWorld/startup/HelloWorldApp.jsx].each do |file|
-             copy_file(base_path + file, file)
+             copy_or_template(base_path + file, convert_filename_to_use_example_page_name(file))
            end
       end
 
@@ -31,11 +34,13 @@ module ReactOnRails
         location = "client/app/bundles/HelloWorld/"
         source = base_path + location
         config = {
-          component_name: "HelloWorldApp",
-          app_relative_path: "./HelloWorldApp"
+          component_name: "#{example_page_name}App",
+          app_relative_path: "./#{example_page_name}App"
         }
-        template("#{source}/startup/registration.jsx.tt", "#{location}/startup/registration.jsx", config)
-        template("#{base_path}app/views/hello_world/index.html.erb.tt", "app/views/hello_world/index.html.erb", config)
+        template("#{source}/startup/registration.jsx.tt",
+                 convert_filename_to_use_example_page_name("#{location}/startup/registration.jsx"), config)
+        template("#{base_path}app/views/hello_world/index.html.erb.tt",
+                 convert_filename_to_use_example_page_name("app/views/hello_world/index.html.erb"), config)
       end
     end
   end