+++ spec/react_on_rails/support/shared_examples/base_generator_examples.rb
@@ -4,7 +4,7 @@ shared_examples "base_generator" do
   it "adds a route for get 'hello_world' to 'hello_world#index'" do
     match = <<-MATCH.strip_heredoc
       Rails.application.routes.draw do
-        get 'hello_world', to: 'hello_world#index'
+        get '#{example_page_path}', to: '#{example_page_path}#index'
       end
     MATCH
     assert_file "config/routes.rb", match
@@ -25,20 +25,22 @@ shared_examples "base_generator" do
 
   it "creates react directories" do
     dirs = %w[components containers startup]
-    dirs.each { |dirname| assert_directory "client/app/bundles/HelloWorld/#{dirname}" }
+    dirs.each do |dirname|
+      assert_directory "client/app/bundles/#{example_page_name}/#{dirname}"
+    end
   end
 
   it "copies react files" do
-    %w[app/controllers/hello_world_controller.rb
-       client/app/bundles/HelloWorld/components/HelloWorld.jsx
-       client/REACT_ON_RAILS_CLIENT_README.md
-       client/webpack.config.js
-       client/.babelrc
-       client/package.json
-       config/initializers/react_on_rails.rb
-       config/webpacker_lite.yml
-       package.json
-       Procfile.dev].each { |file| assert_file(file) }
+    ["app/controllers/#{example_page_path}_controller.rb",
+     "client/app/bundles/#{example_page_name}/components/#{example_page_name}.jsx",
+     "client/REACT_ON_RAILS_CLIENT_README.md",
+     "client/webpack.config.js",
+     "client/.babelrc",
+     "client/package.json",
+     "config/initializers/react_on_rails.rb",
+     "config/webpacker_lite.yml",
+     "package.json",
+     "Procfile.dev"].each { |file| assert_file(file) }
   end
 
   it "templates HelloWorldApp into webpack.config.js" do