+++ spec/react_on_rails/support/shared_examples/react_no_redux_generator_examples.rb
@@ -2,11 +2,11 @@
 
 shared_examples "no_redux_generator" do
   it "creates appropriate templates" do
-    assert_file("client/app/bundles/HelloWorld/startup/registration.jsx") do |contents|
-      assert_match("import HelloWorld from '../components/HelloWorld';", contents)
+    assert_file("client/app/bundles/#{example_page_name}/startup/registration.jsx") do |contents|
+      assert_match("import #{example_page_name} from '../components/#{example_page_name}';", contents)
     end
-    assert_file("app/views/hello_world/index.html.erb") do |contents|
-      assert_match(/"HelloWorld"/, contents)
+    assert_file("app/views/#{example_page_path}/index.html.erb") do |contents|
+      assert_match(/"#{example_page_name}"/, contents)
     end
   end