+++ spec/react_on_rails/support/shared_examples/react_with_redux_generator_examples.rb
@@ -2,15 +2,17 @@
 
 shared_examples "react_with_redux_generator" do
   it "creates redux directories" do
-    %w[actions constants reducers store].each { |dir| assert_directory("client/app/bundles/HelloWorld/#{dir}") }
+    %w[actions constants reducers store].each do |dir|
+      assert_directory("client/app/bundles/#{example_page_name}/#{dir}")
+    end
   end
 
   it "creates appropriate templates" do
-    assert_file("client/app/bundles/HelloWorld/startup/registration.jsx") do |contents|
-      assert_match("import HelloWorldApp from './HelloWorldApp';", contents)
+    assert_file("client/app/bundles/#{example_page_name}/startup/registration.jsx") do |contents|
+      assert_match("import #{example_page_name}App from './#{example_page_name}App';", contents)
     end
-    assert_file("app/views/hello_world/index.html.erb") do |contents|
-      assert_match(/"HelloWorldApp"/, contents)
+    assert_file("app/views/#{example_page_path}/index.html.erb") do |contents|
+      assert_match(/"#{example_page_name}App"/, contents)
     end
   end
 
@@ -20,6 +22,9 @@ shared_examples "react_with_redux_generator" do
        client/app/bundles/HelloWorld/constants/helloWorldConstants.jsx
        client/app/bundles/HelloWorld/reducers/helloWorldReducer.jsx
        client/app/bundles/HelloWorld/store/helloWorldStore.jsx
-       client/app/bundles/HelloWorld/startup/HelloWorldApp.jsx].each { |file| assert_file(file) }
+       client/app/bundles/HelloWorld/startup/HelloWorldApp.jsx].each do |file|
+      file_using_passed_example_page_name = convert_example_page_name(file)
+      assert_file(file_using_passed_example_page_name)
+    end
   end
 end